requirements.txt
fastapi
uvicorn
mlflow
scikit-learn
pandas
numpy
joblib
python-dotenv
dvc[s3]
boto3
pydantic
